Add default Python entry point for module
