+++ src/cryptoserve/server.py
@@ -1,37 +1,8 @@
-import argparse
 import asyncio
-from typing import Any, Callable, Optional
+from typing import Callable
 
 from cryptoserve.greeting import EXERCISES, GREETING
 from cryptoserve.messaging import Client
-
-
-class ArgparseFormatter(
-    argparse.ArgumentDefaultsHelpFormatter, argparse.MetavarTypeHelpFormatter
-):
-    pass
-
-
-def main():
-    argparser = argparse.ArgumentParser(
-        prog="Cryptoserve",
-        description="Server software that houses a library of cryptography-related learning exercises.",
-        formatter_class=ArgparseFormatter,
-    )
-    argparser.add_argument(
-        "--host", type=str, default="127.0.0.1", help="host address to bind to"
-    )
-
-    argparser.add_argument(
-        "--port", "-p", type=int, default=5050, help="port number to bind to"
-    )
-
-    args = argparser.parse_args()
-
-    if not (0 <= args.port <= 65535):
-        argparser.error("port must be in range 0-65535")
-
-    asyncio.run(serve(args.host, args.port))
 
 
 async def serve(host: str, port: int):
@@ -71,7 +42,3 @@ async def run_exercise(client: Client, exercise: Callable[[Client], None]):
 
 def clamp(x, minimum, maximum):
     return max(minimum, min(x, maximum))
-
-
-if __name__ == "__main__":
-    asyncio.run(main())
